+++ tests/test_cuttlepool.py
@@ -8,7 +8,8 @@ import time
 
 import pytest
 
-from cuttlepool import CuttlePool, Resource, _ResourceTracker, PoolDepletedError
+from cuttlepool import (_ResourceTracker, CuttlePool, Resource, PoolEmptyError,
+                        PoolFullError)
 import mockresource
 
 
@@ -78,6 +79,13 @@ def test_resource_wrapper():
     assert isinstance(r, SubResource)
 
 
+def test_empty(pool):
+    """Tests if pool is empty."""
+    assert pool.empty
+    pool.get_resource().close()
+    assert not pool.empty
+
+
 def test_resource_wrapper_get_resource():
     """
     Tests the proper Resource subclass is returned from ``get_resource()``.
@@ -87,19 +95,38 @@ def test_resource_wrapper_get_resource():
     assert isinstance(r, SubResource)
 
 
+def test_get_empty(pool):
+    """Tests the pool raises a ``PoolEmptyError``."""
+    with pytest.raises(PoolEmptyError):
+        pool._get(0)
+
+
+def test_get(pool, resource):
+    """Tests ``_get()`` gets a resource."""
+    resource.close()            # Returns resource to pool.
+    rt = pool._get(0)
+    assert isinstance(rt, _ResourceTracker)
+
+
+def test_get_wait():
+    def worker(r):
+        time.sleep(5)
+        r.close()
+
+    pool = MockPool(mockresource.factory, capacity=1)
+    resource = pool.get_resource()
+
+    t = threading.Thread(target=worker, args=(resource,))
+    t.start()
+
+    rt = pool._get(None)
+    assert isinstance(rt, _ResourceTracker)
+
+
 def test_get_tracker(pool, rtracker):
     """Tests the resource tracker for a resource is returned."""
     rt = pool._get_tracker(rtracker.resource)
     assert rt is rtracker
-
-
-def test_make_resource(pool):
-    """
-    Tests the resource object returned from _make_resource is the proper class
-    instance.
-    """
-    r = pool._make_resource()
-    assert isinstance(r, _ResourceTracker)
 
 
 def test_harvest_lost_resources(pool):
@@ -118,6 +145,42 @@ def test_harvest_lost_resources(pool):
     assert r_id == id(pool.get_resource()._resource)
 
 
+def test_make_resource(pool):
+    """
+    Tests the resource object returned from _make_resource is the proper class
+    instance.
+    """
+    r = pool._make_resource()
+    assert pool.size == 1
+    assert isinstance(r, _ResourceTracker)
+
+
+def test_put_full():
+    """Tests ``PoolFullError`` is raised."""
+    pool = MockPool(mockresource.factory, capacity=1, overflow=1)
+    r1 = pool.get_resource()
+    r2 = pool.get_resource()
+    print(pool.capacity, pool._available)
+
+    pool._put(pool._get_tracker(r1._resource))
+    with pytest.raises(PoolFullError):
+        pool._put(pool._get_tracker(r2._resource))
+
+
+def test_put(pool, rtracker):
+    """Tests ``_put()`` returns resource to pool."""
+    assert pool._available == 0
+    pool._put(rtracker)
+    assert pool._available == 1
+
+
+def test_remove(pool, rtracker):
+    """Tests ``_remove()`` removes resource from pool."""
+    pool._remove(rtracker)
+    assert pool.size == pool._available == 0
+    assert list(filter(None, pool._reference_queue)) == []
+
+
 def test_get_resource(pool):
     """
     Tests the resource object returned from get_resource is the
@@ -131,18 +194,16 @@ def test_get_resource_overflow(pool):
     """
     Tests the pool creates proper number of overflow resources properly.
     """
-    rt = []
-    for __ in range(pool._capacity):
-        rt.append(pool.get_resource())
+    rs = []
+    for _ in range(pool.maxsize):
+        rs.append(pool.get_resource())
 
-    r = pool.get_resource()
     assert pool.size == pool.maxsize
 
-    r.close()
-    for r in rt:
+    for r in rs:
         r.close()
 
-    assert pool.size == pool._capacity
+    assert pool.size == pool.capacity
 
 
 def test_get_resource_depleted(pool):
@@ -163,7 +224,7 @@ def test_get_resource_depleted(pool):
 def test_get_resource_depleted_error():
     """Tests the pool will raise an error when depleted."""
     pool = MockPool(mockresource.factory, capacity=1, timeout=1)
-    with pytest.raises(PoolDepletedError):
+    with pytest.raises(PoolEmptyError):
         rt = []
         while True:
             rt.append(pool.get_resource())
@@ -273,3 +334,15 @@ def test_close(pool):
     r.close()
     assert r._resource is None
     assert r._pool is None
+
+
+def test_recycling(pool):
+    """
+    Tests no errors are raised for multiple rounds of getting and putting.
+    """
+    for _ in range(5):
+        rs = [pool.get_resource() for _ in range(pool.maxsize)]
+        for r in rs:
+            r.close()
+
+    assert pool._available == pool.size == pool.capacity